netmonitor/requirements.txt
scapy
dnspython
